+++ System/UIObject.java
@@ -5,8 +5,8 @@ interface UIObject {
     abstract void Update();
     abstract void setSize(int width, int height);
     abstract void setSize(float size);
-    abstract void setPosition(int x, int y);
-    abstract void move(int x, int y);
+    abstract void setPosition(float x, float y);
+    abstract void move(float x, float y);
     abstract int getWidth();
     abstract int getHeight();
     abstract JLabel getObject();
